backend/requirements.txt
uvicorn[standard]==0.30.0
fastapi==0.111.0
pydantic==2.7.1
websockets==12.0
httpx==0.27.0
pandas==2.2.2
numpy==1.26.4
ta==0.11.0
yfinance==0.2.40
torch==2.3.0
scikit-learn==1.4.2
python-dotenv==1.0.1
openai==1.51.0
alpaca-py==0.25.0
newsapi-python==0.2.7